app/src/androidTest/java/uk/gov/onelogin/login/LoadingScreenTest.kt
package uk.gov.onelogin.login

import androidx.compose.ui.test.assertIsDisplayed
import androidx.compose.ui.test.junit4.createComposeRule
import androidx.compose.ui.test.onNodeWithTag
import dagger.hilt.android.testing.HiltAndroidTest
import org.junit.Before
import org.junit.Rule
import org.junit.Test
import uk.gov.onelogin.TestCase
import uk.gov.onelogin.ext.setupComposeTestRule

@HiltAndroidTest
class LoadingScreenTest : TestCase() {
    @get:Rule
    val composeTestRule = createComposeRule()

    @Before
    fun setupNavigation() {
        composeTestRule.setupComposeTestRule { _ ->
            LoadingScreen()
        }
    }

    @Test
    fun verifyComponents() {
        composeTestRule.onNodeWithTag(LOADING_SCREEN_BOX).assertIsDisplayed()
        composeTestRule.onNodeWithTag(LOADING_SCREEN_PROGRESS_INDICATOR).assertIsDisplayed()
    }
}
